+++ glue/core/layer_artist.py
@@ -83,6 +83,13 @@ class LayerArtistBase(PropertySetMixin):
 
         self._disabled_reason = ''  # A string explaining why this layer is disabled.
 
+    def enable(self):
+        if self.enabled:
+            return
+        self._disabled_reason = ''
+        self._enabled = True
+        self.redraw()
+
     def disable(self, reason):
         """
         Disable the layer for a particular reason.