+++ glue/core/link_manager.py
@@ -27,6 +27,11 @@ from glue.core.link_helpers import LinkCollection
 from glue.core.component_link import ComponentLink
 from glue.core.data import Data
 from glue.core.component import DerivedComponent
+from glue.core.exceptions import IncompatibleAttribute
+
+
+__all__ = ['accessible_links', 'discover_links', 'find_dependents',
+           'LinkManager', 'is_equivalent_cid']
 
 
 def accessible_links(cids, links):
@@ -42,8 +47,7 @@ def accessible_links(cids, links):
     """
     cids = set(cids)
     return [l for l in links if
-             set(l.get_from_ids()) <= cids]
-
+            set(l.get_from_ids()) <= cids]
 
 
 def discover_links(data, links):
@@ -126,10 +130,12 @@ class LinkManager(HubListener):
 
     def __init__(self):
         self._links = set()
+        self.hub = None
 
     def register_to_hub(self, hub):
-        hub.subscribe(self, DataCollectionDeleteMessage,
-                      handler=self._data_removed)
+        self.hub = hub
+        self.hub.subscribe(self, DataCollectionDeleteMessage,
+                           handler=self._data_removed)
 
     def _data_removed(self, msg):
         remove = []
@@ -155,13 +161,17 @@ class LinkManager(HubListener):
             for l in link:
                 self.add_link(l)
         else:
-            if not link.inverse in self._links:
+            if link.inverse not in self._links:
                 self._links.add(link)
 
     @contract(link=ComponentLink)
     def remove_link(self, link):
-        logging.getLogger(__name__).debug('removing link %s', link)
-        self._links.remove(link)
+        if isinstance(link, (LinkCollection, list)):
+            for l in link:
+                self.remove_link(l)
+        else:
+            logging.getLogger(__name__).debug('removing link %s', link)
+            self._links.remove(link)
 
     @contract(data=Data)
     def update_data_components(self, data):
@@ -200,8 +210,13 @@ class LinkManager(HubListener):
         for m in missing_links:
             to_remove.extend(find_dependents(data, m))
 
-        for r in to_remove:
-            data.remove_component(r)
+        if getattr(data, 'hub', None) is None:
+            for r in to_remove:
+                data.remove_component(r)
+        else:
+            with data.hub.delay_callbacks():
+                for r in to_remove:
+                    data.remove_component(r)
 
     def _add_deriveable_components(self, data):
         """Find and add any DerivedComponents that a data object can
@@ -224,3 +239,43 @@ class LinkManager(HubListener):
 
     def __contains__(self, item):
         return item in self._links
+
+
+def _find_identical_reference_cid(data, cid):
+    """
+    Given a dataset and a component ID, return the equivalent component ID that
+    truly belongs to the dataset (not via a link). Returns None if there is
+    no strictly identical component in the dataset.
+    """
+    try:
+        target_comp = data.get_component(cid)
+    except IncompatibleAttribute:
+        return None
+    if isinstance(target_comp, DerivedComponent):
+        if target_comp.link.identity:
+            updated_cid = target_comp.link.get_from_ids()[0]
+            return _find_identical_reference_cid(data, updated_cid)
+        else:
+            return None
+    else:
+        return cid
+
+
+def is_equivalent_cid(data, cid1, cid2):
+    """
+    Convenience function to determine if two component IDs in a dataset are
+    equivalent.
+
+    Parameters
+    ----------
+    data : `~glue.core.Data`
+        The data object in which to check for the component IDs
+    cid1, cid2 : `~glue.core.ComponentID`
+        The two component IDs to compare
+    """
+
+    # Dereference the component IDs to find base component ID
+    cid1 = _find_identical_reference_cid(data, cid1)
+    cid2 = _find_identical_reference_cid(data, cid2)
+
+    return cid1 is cid2